+++ python/ray/rllib/test/test_filters.py
@@ -5,7 +5,10 @@ from __future__ import print_function
 import unittest
 import numpy as np
 
+import ray
 from ray.rllib.utils.filter import RunningStat, MeanStdFilter
+from ray.rllib.utils import FilterManager
+from ray.rllib.test.mock import _MockEvaluator
 
 
 class RunningStatTest(unittest.TestCase):
@@ -68,6 +71,37 @@ class MSFTest(unittest.TestCase):
             filt.apply_changes(filt2, with_buffer=True)
             self.assertEqual(filt.buffer.n, 5)
             self.assertEqual(filt.rs.n, 15)
+
+
+class FilterManagerTest(unittest.TestCase):
+    def setUp(self):
+        ray.init(num_cpus=1)
+
+    def tearDown(self):
+        ray.worker.cleanup()
+
+    def testSynchronize(self):
+        """Synchronize applies filter buffer onto own filter"""
+        filt1 = MeanStdFilter(())
+        for i in range(10):
+            filt1(i)
+        self.assertEqual(filt1.rs.n, 10)
+        filt1.clear_buffer()
+        self.assertEqual(filt1.buffer.n, 0)
+
+        RemoteEvaluator = ray.remote(_MockEvaluator)
+        remote_e = RemoteEvaluator.remote(sample_count=10)
+        remote_e.sample.remote()
+
+        manager = FilterManager(obs_filter=filt1)
+        manager.synchronize([remote_e])
+
+        filters = ray.get(remote_e.get_filters.remote())
+        obs_f = filters["obs_filter"]
+        self.assertEqual(filt1.rs.n, 20)
+        self.assertEqual(filt1.buffer.n, 0)
+        self.assertEqual(obs_f.rs.n, filt1.rs.n)
+        self.assertEqual(obs_f.buffer.n, filt1.buffer.n)
 
 
 if __name__ == "__main__":
